+++ src/main/java/io/github/astrapi69/crypto/factories/CertFactory.java
@@ -44,6 +44,7 @@ import java.util.Date;
 
 import javax.security.auth.x500.X500Principal;
 
+import io.github.astrapi69.throwable.RuntimeExceptionDecorator;
 import org.bouncycastle.asn1.x500.X500Name;
 import org.bouncycastle.asn1.x509.BasicConstraints;
 import org.bouncycastle.asn1.x509.Extension;
@@ -57,8 +58,6 @@ import org.bouncycastle.operator.ContentSigner;
 import org.bouncycastle.operator.OperatorCreationException;
 import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder;
 import org.bouncycastle.x509.X509V3CertificateGenerator;
-
-import com.rainerhahnekamp.sneakythrow.Sneaky;
 
 import io.github.astrapi69.crypto.provider.SecurityProvider;
 
@@ -341,7 +340,8 @@ public final class CertFactory
 		if (extensions != null && 0 < extensions.length)
 		{
 			Arrays.stream(extensions)
-				.forEach(extension -> Sneaky.sneaked(() -> certBuilder.addExtension(extension)));
+				.forEach(RuntimeExceptionDecorator
+					.decorate(extension -> certBuilder.addExtension(extension)));
 		}
 		ContentSigner signer = new JcaContentSignerBuilder(signatureAlgorithm)
 			.setProvider(SecurityProvider.BC.name()).build(keyPair.getPrivate());